Documentation, class weights fix, some matching analysis.
Added significant documentation to matching.py and the notebook: function docstrings, more comments, more explanation in the notebook about alternative methods considered.
Changed the weights per class to be proportional to the number of tutees requesting help in that class divided by the number of tutors available for that class. This prevents the priorities from being game-able by a student requesting an absurd number of hours (which would give the classes they requested much higher priority under the old scheme) and also re-weights by the number of tutors available, so that the priority is more to classes that are difficult to find tutoring in.
Also added a function in the notebook to characterize matchings by comparing the number of tutees, matched tutees, tutors, matched tutors, etc. between differing matchings files.