Implemented iterative matching to allow for some non-linear constraints.
The old approach is still usable (switch to it with a command line flag), but the default is now to match each tutor to at most one tutee and each tutee to at most one tutor in the LP section. This is done iteratively until no more matches can be made. The constraints are now enforced that tutors can have at most 3 tutees and that tutees can have at most 1 tutor per class (having multiple tutors per class was the major problem that led to this approach).

The number of hours assigned to a matching is also now set based on a rule system. The LP portion only assigns a single hour to each matching, but once a tutor cannot be matched with any new tutees, any time they have left available is split amongst their tutees (the non-iterative LP is used to do this part).

We also now swap tutors with 0 matches in for tutors with multiple matches where possible.

More informative file names are used which specify the values for lambda classes, lambda students, and whether iterative or single matching was used; previously the file names just had an integer appended.